feat: use php scoper

+++ vendor-prefixed/symfony/var-dumper/Dumper/AbstractDumper.php
@@ -7,15 +7,11 @@
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
- *
- * Modified by hirasso on 25-December-2024 using {@see https://github.com/BrianHenryIE/strauss}.
  */
-
 namespace RH\AdminUtils\Symfony\Component\VarDumper\Dumper;
 
 use RH\AdminUtils\Symfony\Component\VarDumper\Cloner\Data;
 use RH\AdminUtils\Symfony\Component\VarDumper\Cloner\DumperInterface;
-
 /**
  * Abstract mechanism for dumping a Data object.
  *
@@ -27,10 +23,8 @@ abstract class AbstractDumper implements DataDumperInterface, DumperInterface
     public const DUMP_STRING_LENGTH = 2;
     public const DUMP_COMMA_SEPARATOR = 4;
     public const DUMP_TRAILING_COMMA = 8;
-
     /** @var callable|resource|string|null */
     public static $defaultOutput = 'php://output';
-
     protected string $line = '';
     /** @var callable|null */
     protected $lineDumper;
@@ -38,25 +32,20 @@ abstract class AbstractDumper implements DataDumperInterface, DumperInterface
     protected $outputStream;
     protected string $decimalPoint = '.';
     protected string $indentPad = '  ';
-    protected int $flags;
-
     private string $charset = '';
-
     /**
      * @param callable|resource|string|null $output  A line dumper callable, an opened stream or an output path, defaults to static::$defaultOutput
      * @param string|null                   $charset The default character encoding to use for non-UTF8 strings
      * @param int                           $flags   A bit field of static::DUMP_* constants to fine tune dumps representation
      */
-    public function __construct($output = null, ?string $charset = null, int $flags = 0)
+    public function __construct($output = null, ?string $charset = null, protected int $flags = 0)
     {
-        $this->flags = $flags;
-        $this->setCharset($charset ?: \ini_get('php.output_encoding') ?: \ini_get('default_charset') ?: 'UTF-8');
+        $this->setCharset((($charset ?: \ini_get('php.output_encoding')) ?: \ini_get('default_charset')) ?: 'UTF-8');
         $this->setOutput($output ?: static::$defaultOutput);
         if (!$output && \is_string(static::$defaultOutput)) {
             static::$defaultOutput = $this->outputStream;
         }
     }
-
     /**
      * Sets the output destination of the dumps.
      *
@@ -67,7 +56,6 @@ abstract class AbstractDumper implements DataDumperInterface, DumperInterface
     public function setOutput($output)
     {
         $prev = $this->outputStream ?? $this->lineDumper;
-
         if (\is_callable($output)) {
             $this->outputStream = null;
             $this->lineDumper = $output;
@@ -78,10 +66,8 @@ abstract class AbstractDumper implements DataDumperInterface, DumperInterface
             $this->outputStream = $output;
             $this->lineDumper = $this->echoLine(...);
         }
-
         return $prev;
     }
-
     /**
      * Sets the default character encoding to use for non-UTF8 strings.
      *
@@ -90,15 +76,11 @@ abstract class AbstractDumper implements DataDumperInterface, DumperInterface
     public function setCharset(string $charset): string
     {
         $prev = $this->charset;
-
         $charset = strtoupper($charset);
-        $charset = 'UTF-8' === $charset || 'UTF8' === $charset ? 'CP1252' : $charset;
-
+        $charset = ('UTF-8' === $charset || 'UTF8' === $charset) ? 'CP1252' : $charset;
         $this->charset = $charset;
-
         return $prev;
     }
-
     /**
      * Sets the indentation pad string.
      *
@@ -110,10 +92,8 @@ abstract class AbstractDumper implements DataDumperInterface, DumperInterface
     {
         $prev = $this->indentPad;
         $this->indentPad = $pad;
-
         return $prev;
     }
-
     /**
      * Dumps a Data object.
      *
@@ -123,11 +103,10 @@ abstract class AbstractDumper implements DataDumperInterface, DumperInterface
      */
     public function dump(Data $data, $output = null): ?string
     {
-        if ($locale = $this->flags & (self::DUMP_COMMA_SEPARATOR | self::DUMP_TRAILING_COMMA) ? setlocale(\LC_NUMERIC, 0) : null) {
+        if ($locale = ($this->flags & (self::DUMP_COMMA_SEPARATOR | self::DUMP_TRAILING_COMMA)) ? setlocale(\LC_NUMERIC, 0) : null) {
             setlocale(\LC_NUMERIC, 'C');
         }
-
-        if ($returnDump = true === $output) {
+        if ($returnDump = \true === $output) {
             $output = fopen('php://memory', 'r+');
         }
         if ($output) {
@@ -136,11 +115,9 @@ abstract class AbstractDumper implements DataDumperInterface, DumperInterface
         try {
             $data->dump($this);
             $this->dumpLine(-1);
-
             if ($returnDump) {
                 $result = stream_get_contents($output, -1, 0);
                 fclose($output);
-
                 return $result;
             }
         } finally {
@@ -151,10 +128,8 @@ abstract class AbstractDumper implements DataDumperInterface, DumperInterface
                 setlocale(\LC_NUMERIC, $locale);
             }
         }
-
         return null;
     }
-
     /**
      * Dumps the current line.
      *
@@ -166,17 +141,15 @@ abstract class AbstractDumper implements DataDumperInterface, DumperInterface
         ($this->lineDumper)($this->line, $depth, $this->indentPad);
         $this->line = '';
     }
-
     /**
      * Generic line dumper callback.
      */
     protected function echoLine(string $line, int $depth, string $indentPad): void
     {
         if (-1 !== $depth) {
-            fwrite($this->outputStream, str_repeat($indentPad, $depth).$line."\n");
+            fwrite($this->outputStream, str_repeat($indentPad, $depth) . $line . "\n");
         }
     }
-
     /**
      * Converts a non-UTF-8 string to UTF-8.
      */
@@ -185,18 +158,15 @@ abstract class AbstractDumper implements DataDumperInterface, DumperInterface
         if (null === $s || preg_match('//u', $s)) {
             return $s;
         }
-
-        if (!\function_exists('iconv')) {
+        if (!\function_exists('iconv') && !\function_exists('RH\AdminUtils\iconv')) {
             throw new \RuntimeException('Unable to convert a non-UTF-8 string to UTF-8: required function iconv() does not exist. You should install ext-iconv or symfony/polyfill-iconv.');
         }
-
-        if (false !== $c = @iconv($this->charset, 'UTF-8', $s)) {
+        if (\false !== $c = @iconv($this->charset, 'UTF-8', $s)) {
             return $c;
         }
-        if ('CP1252' !== $this->charset && false !== $c = @iconv('CP1252', 'UTF-8', $s)) {
+        if ('CP1252' !== $this->charset && \false !== $c = @iconv('CP1252', 'UTF-8', $s)) {
             return $c;
         }
-
         return iconv('CP850', 'UTF-8', $s);
     }
 }
